+++ internal/impl/aws/session/type.go
@@ -2,12 +2,13 @@ package session
 
 // CredentialsConfig contains configuration params for AWS credentials.
 type CredentialsConfig struct {
-	Profile    string `json:"profile" yaml:"profile"`
-	ID         string `json:"id" yaml:"id"`
-	Secret     string `json:"secret" yaml:"secret"`
-	Token      string `json:"token" yaml:"token"`
-	Role       string `json:"role" yaml:"role"`
-	ExternalID string `json:"role_external_id" yaml:"role_external_id"`
+	Profile     string `json:"profile" yaml:"profile"`
+	ID          string `json:"id" yaml:"id"`
+	Secret      string `json:"secret" yaml:"secret"`
+	Token       string `json:"token" yaml:"token"`
+	UseEC2Creds bool   `json:"from_ec2_role" yaml:"from_ec2_role"`
+	Role        string `json:"role" yaml:"role"`
+	ExternalID  string `json:"role_external_id" yaml:"role_external_id"`
 }
 
 // Config contains configuration fields for an AWS session. This config is