+++ internal/impl/elasticsearch/output.go
@@ -51,10 +51,10 @@ false for connections to succeed.`),
 		Config: docs.FieldComponent().WithChildren(
 			docs.FieldString("urls", "A list of URLs to connect to. If an item of the list contains commas it will be expanded into multiple URLs.", []string{"http://localhost:9200"}).Array(),
 			docs.FieldString("index", "The index to place messages.").IsInterpolated(),
-			docs.FieldString("action", "The action to take on the document.").IsInterpolated().HasOptions("index", "update", "delete").Advanced(),
+			docs.FieldString("action", "The action to take on the document.").IsInterpolated().HasOptions("create", "index", "update", "delete").Advanced(),
 			docs.FieldString("pipeline", "An optional pipeline id to preprocess incoming documents.").IsInterpolated().Advanced(),
 			docs.FieldString("id", "The ID for indexed messages. Interpolation should be used in order to create a unique ID for each message.").IsInterpolated(),
-			docs.FieldString("type", "The document type.").Deprecated(),
+			docs.FieldString("type", "The document mapping type. This field is required for versions of elasticsearch earlier than 6.0.0, but are invalid for versions 7.0.0 or later.").Optional(),
 			docs.FieldString("routing", "The routing key to use for the document.").IsInterpolated().Advanced(),
 			docs.FieldBool("sniff", "Prompts Benthos to sniff for brokers to connect to when establishing a connection.").Advanced(),
 			docs.FieldBool("healthcheck", "Whether to enable healthchecks.").Advanced(),
@@ -383,6 +383,17 @@ func (e *Elasticsearch) buildBulkableRequest(p *pendingBulkIndex) (elastic.Bulka
 		return r, nil
 	case "index":
 		r := elastic.NewBulkIndexRequest().
+			Index(p.Index).
+			Pipeline(p.Pipeline).
+			Routing(p.Routing).
+			Id(p.ID).
+			Doc(p.Doc)
+		if p.Type != "" {
+			r = r.Type(p.Type)
+		}
+		return r, nil
+	case "create":
+		r := elastic.NewBulkCreateRequest().
 			Index(p.Index).
 			Pipeline(p.Pipeline).
 			Routing(p.Routing).