+++ public/bloblang/spec.go
@@ -82,11 +82,14 @@ func (d ParamDefinition) Default(v interface{}) ParamDefinition {
 // following the order in which the parameters are added, or named style
 // (c: baz, a: foo).
 type PluginSpec struct {
+	status      query.Status
 	category    string
 	description string
 	impure      bool
+	static      bool
 	params      query.Params
 	examples    []pluginExample
+	version     string
 }
 
 type pluginExample struct {
@@ -103,6 +106,25 @@ func NewPluginSpec() *PluginSpec {
 	}
 }
 
+// Experimental flags the plugin as an experimental component.
+func (p *PluginSpec) Experimental() *PluginSpec {
+	p.status = query.StatusExperimental
+	return p
+}
+
+// Beta flags the plugin as a beta component.
+func (p *PluginSpec) Beta() *PluginSpec {
+	p.status = query.StatusBeta
+	return p
+}
+
+// Deprecated flags the plugin as a deprecated component, it will still be valid
+// in mappings but won't appear prominently in documentation.
+func (p *PluginSpec) Deprecated() *PluginSpec {
+	p.status = query.StatusDeprecated
+	return p
+}
+
 // Category adds an optional category string to the plugin spec, this is used
 // when generating documentation for the plugin.
 func (p *PluginSpec) Category(str string) *PluginSpec {
@@ -114,6 +136,12 @@ func (p *PluginSpec) Category(str string) *PluginSpec {
 // when generating documentation for the plugin.
 func (p *PluginSpec) Description(str string) *PluginSpec {
 	p.description = str
+	return p
+}
+
+// Version specifies that this plugin was introduced in a given version.
+func (p *PluginSpec) Version(v string) *PluginSpec {
+	p.version = v
 	return p
 }
 
@@ -144,6 +172,18 @@ func (p *PluginSpec) Param(def ParamDefinition) *PluginSpec {
 // are excluded from some bloblang environments.
 func (p *PluginSpec) Impure() *PluginSpec {
 	p.impure = true
+	return p
+}
+
+// Static marks the plugin as a statically evaluated function or method. This is
+// a guarantee that given the name parameters this plugin will always yield the
+// same value.
+//
+// Marking a function or method as static has the advantage that it can
+// sometimes be optimistically evaluated at mapping parse time when given static
+// arguments.
+func (p *PluginSpec) Static() *PluginSpec {
+	p.static = true
 	return p
 }